+++ changelog.md
@@ -18,5 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a json controller
 - Add a dto package
 - The tasks in the CI are ran in parallel.
+- Add a Generic CRUD Repository
 
 [unreleased]: https://github.com/ditrit/badaas/blob/main/changelog.md#unreleased